[20기_이한슬] CRUD API 미션 제출합니다.

+++ src/main/java/com/ceos20/instagram_clone/domain/comment/service/CommentService.java
@@ -1,7 +1,7 @@
 package com.ceos20.instagram_clone.domain.comment.service;
 
-import com.ceos20.instagram_clone.domain.comment.dto.request.CommentReq;
-import com.ceos20.instagram_clone.domain.comment.dto.response.CommentRes;
+import com.ceos20.instagram_clone.domain.comment.dto.request.CommentRequestDto;
+import com.ceos20.instagram_clone.domain.comment.dto.response.CommentResponseDto;
 import com.ceos20.instagram_clone.domain.comment.entity.Comment;
 import com.ceos20.instagram_clone.domain.member.entity.Member;
 import com.ceos20.instagram_clone.domain.post.entity.Post;
@@ -29,30 +29,29 @@ public class CommentService {
     private final MemberRepository memberRepository;
 
     public Member findMemberById(Long memberId) {
-        return memberRepository.findById(memberId).orElseThrow(() -> new BadRequestException(NOT_FOUND_MEMBER_ID));
+        return memberRepository.findByIdAndDeletedAtIsNull(memberId).orElseThrow(() -> new BadRequestException(NOT_FOUND_MEMBER_ID));
     }
 
     public Post findPostById(Long postId) {
-        return postRepository.findById(postId).orElseThrow(() -> new BadRequestException(NOT_FOUND_POST_ID));
+        return postRepository.findByIdAndDeletedAtIsNull(postId).orElseThrow(() -> new BadRequestException(NOT_FOUND_POST_ID));
     }
 
     public Comment findCommentById(Long commentId) {
-        return commentRepository.findById(commentId).orElseThrow(() -> new BadRequestException(NOT_FOUND_COMMENT_ID));
+        return commentRepository.findByIdAndDeletedAtIsNull(commentId).orElseThrow(() -> new BadRequestException(NOT_FOUND_COMMENT_ID));
     }
 
-    /** [주요 기능 ]
+    /** [주요 기능]
      * 댓글 생성
      * **/
     @Transactional
-    public CommentRes createComment(CommentReq request) {
+    public CommentResponseDto createComment(CommentRequestDto request, Long postId, Long memberId) {
 
-        Post post = findPostById(request.postId());
-        Member member = findMemberById(request.memberId());
+        Post post = findPostById(postId);
+        Member member = findMemberById(memberId);
 
         Comment parentComment = null;
         if (request.parentCommentId() != null) {
-            parentComment = commentRepository.findById(request.parentCommentId())
-                    .orElseThrow(() -> new BadRequestException(NOT_FOUND_PARENT_COMMENT_ID));
+            parentComment = findCommentById(request.parentCommentId());
         }
 
         Comment comment = request.toEntity(post, member, parentComment);
@@ -61,33 +60,44 @@ public class CommentService {
         post.updateCommentCount(post.getCommentCount() + 1);
         postRepository.save(post);
 
-        return CommentRes.createCommentRes(saveComment);
+        return CommentResponseDto.from(saveComment);
     }
 
     /** [주요 기능 ]
      * 댓글 조회
      * **/
-    public CommentRes getComment(Long commentId) {
+    public CommentResponseDto getComment(Long commentId) {
 
         Comment comment = findCommentById(commentId);
 
-        List<CommentRes> replies = comment.getReplies() != null
-                ? comment.getReplies().stream()
-                .map(CommentRes::createCommentRes)
-                .collect(Collectors.toList())
-                : new ArrayList<>();
+        return CommentResponseDto.from(comment);
+    }
 
-        return CommentRes.getCommentRes(comment, replies);
+    /** [주요 기능 ]
+     * 게시물에 있는 전체 댓글 조회
+     * **/
+    public List<CommentResponseDto> getAllPostComment(Long postId) {
+
+        Post post = findPostById(postId);
+        List<Comment> comments = commentRepository.findAllByPostAndParentCommentIsNull(post);
+        return comments.stream()
+                .map(CommentResponseDto::from)
+                .toList();
     }
 
     /** [주요 기능 ]
      * 댓글 삭제
      * **/
     @Transactional
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, Long postId, Long memberId) {
 
         Comment comment = findCommentById(commentId);
-        Post post = comment.getPost();
+        Post post = findPostById(postId);
+        Member member = findMemberById(memberId);
+
+        if (!comment.getMember().equals(member)) {
+            throw new BadRequestException(INVALID_AUTHORITY);
+        }
 
         commentRepository.delete(comment);
